requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
langgraph
langchain-groq
langchain-core
requests
APScheduler
pandas
numpy
scikit-learn
statsmodels
shap
